+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 04/12/2019   
+# 27/02/2020
+
+- Add possibility of sorting datasets by user fields (such as name or role).
+
+# 04/12/2019
 - Allow microservices to update the dataset's `errorMessage` field
 
 # 26/11/2019